requirements.txt
# Django Backend
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# Image Processing
Pillow==10.1.0

# ML/AI
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
onnxruntime==1.16.3
onnxscript>=0.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
transformers==4.35.2
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.18

# FastAPI (for modular services if needed)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
